requirements.txt
absl-py                   2.1.0
addict                    2.4.0
aliyun-python-sdk-core    2.13.36
aliyun-python-sdk-kms     2.16.1
altgraph                  0.17.4
asttokens                 2.4.1
attrs                     23.2.0
blinker                   1.7.0
brotlipy                  0.7.0
certifi                   2023.7.22
cffi                      1.15.1
charset-normalizer        2.0.4
click                     8.1.7
colorama                  0.4.6
comm                      0.2.2
ConfigArgParse            1.7
contourpy                 1.1.0
crcmod                    1.7
cryptography              41.0.2
cycler                    0.11.0
dash                      2.16.1
dash-core-components      2.0.0
dash-html-components      2.0.0
dash-table                5.0.0
decorator                 5.1.1
efficientnet-pytorch      0.7.1
einops                    0.6.1
exceptiongroup            1.2.0
executing                 2.0.1
fastjsonschema            2.19.1
filelock                  3.9.0
Flask                     3.0.2
flatbuffers               24.3.25
fonttools                 4.42.1
fsspec                    2023.9.0
huggingface-hub           0.16.4
idna                      3.4
importlib-metadata        6.8.0
ipython                   8.23.0
ipywidgets                8.1.2
itsdangerous              2.1.2
jax                       0.4.34
jaxlib                    0.4.34
jedi                      0.19.1
Jinja2                    3.1.2
jmespath                  0.10.0
joblib                    1.2.0
jsonschema                4.21.1
jsonschema-specifications 2023.12.1
jupyter_core              5.7.2
jupyterlab_widgets        3.0.10
kiwisolver                1.4.5
lap                       0.4.0
lxml                      4.9.3
Markdown                  3.4.4
markdown-it-py            3.0.0
MarkupSafe                2.1.1
matchers                  0.22
matplotlib                3.8.0rc1
matplotlib-inline         0.1.6
mdurl                     0.1.2
mediapipe                 0.10.14
mkl-fft                   1.3.6
mkl-random                1.2.2
mkl-service               2.4.0
ml_dtypes                 0.5.0
mmcv                      2.0.1
mmengine                  0.8.4
model-index               0.1.11
mpmath                    1.3.0
nbformat                  5.10.4
nest-asyncio              1.6.0
networkx                  3.1
numpy                     1.25.2
opcua                     0.98.13
open3d                    0.18.0
opencv-contrib-python     4.10.0.84
opencv-python             4.8.0.76
opendatalab               0.0.10
openmim                   0.3.9
openxlab                  0.0.24
opt_einsum                3.4.0
ordered-set               4.1.0
oss2                      2.17.0
packaging                 23.1
pandas                    2.1.0
parso                     0.8.4
PCV                       1.0
pefile                    2023.2.7
Pillow                    9.4.0
pip                       23.2.1
platformdirs              3.10.0
plotly                    5.20.0
prompt-toolkit            3.0.43
protobuf                  4.25.5
psutil                    5.9.5
pure-eval                 0.2.2
pycparser                 2.21
pycryptodome              3.18.0
Pygments                  2.16.1
PyHamcrest                2.0.4
pyinstaller               6.1.0
pyinstaller-hooks-contrib 2023.10
pyOpenSSL                 23.2.0
pyparsing                 3.1.1
PyQt5                     5.15.10
PyQt5-Qt5                 5.15.2
PyQt5-sip                 12.13.0
PySocks                   1.7.1
python-dateutil           2.8.2
pytz                      2023.3.post1
pywin32                   306
pywin32-ctypes            0.2.2
PyYAML                    6.0.1
referencing               0.34.0
regex                     2023.8.8
requests                  2.28.2
retrying                  1.3.4
rich                      13.4.2
rpds-py                   0.18.0
safetensors               0.3.3
scikit-learn              1.3.0
scipy                     1.11.1
seaborn                   0.12.2
setuptools                60.2.0
six                       1.16.0
sounddevice               0.5.1
stack-data                0.6.3
sympy                     1.11.1
tabulate                  0.9.0
tenacity                  8.2.3
termcolor                 2.3.0
thop                      0.1.1.post2209072238
threadpoolctl             2.2.0
timm                      0.9.7
tomli                     2.0.1
torch                     2.0.1
torchaudio                2.0.2
torchvision               0.15.2
tqdm                      4.65.2
traitlets                 5.14.2
typing_extensions         4.7.1
tzdata                    2023.3
urllib3                   1.26.16
wcwidth                   0.2.13
Werkzeug                  3.0.2
wheel                     0.38.4
widgetsnbextension        4.0.10
win-inet-pton             1.1.0
yapf                      0.40.1
zipp                      3.16.2
